@@ -9,11 +9,7 @@ class OpenFlutterPromoTile extends StatelessWidget {
   final Function onClickApply;
   final Color textColor;
 
-  const OpenFlutterPromoTile(
-      {Key key,
-      @required this.item,
-      @required this.onClickApply,
-      @required this.textColor})
+  const OpenFlutterPromoTile({Key key, @required this.item, @required this.onClickApply, @required this.textColor})
       : super(key: key);
 
   @override
@@ -35,37 +31,26 @@ class OpenFlutterPromoTile extends StatelessWidget {
         children: <Widget>[
           Container(
               width: 85,
-              decoration: item.image != null ?
-                BoxDecoration(
-                  image: DecorationImage(
-                    fit: BoxFit.fill, image: AssetImage(item.image)
-                  )
-                )
-                : 
-                BoxDecoration(
-                  color: AppColors.red,
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(AppSizes.textFieldRadius),
-                    bottomLeft: Radius.circular(AppSizes.textFieldRadius)),
-                ),
+              decoration: item.image != null
+                  ? BoxDecoration(image: DecorationImage(fit: BoxFit.fill, image: AssetImage(item.image)))
+                  : BoxDecoration(
+                      color: AppColors.red,
+                      borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(AppSizes.textFieldRadius),
+                          bottomLeft: Radius.circular(AppSizes.textFieldRadius)),
+                    ),
               padding: EdgeInsets.all(AppSizes.linePadding * 2),
               child: Row(children: <Widget>[
                 Text(item.discount.toStringAsFixed(0),
-                  style: _theme.textTheme.caption.copyWith(
-                    color: textColor,
-                    fontSize: 32)),
+                    style: _theme.textTheme.caption.copyWith(color: textColor, fontSize: 32)),
                 Padding(
-                  padding: EdgeInsets.only(top: 16, left: 5),
-                  child: Column(
-                    children: <Widget>[
-                      Text('%',
-                        style: _theme.textTheme.display3
-                          .copyWith(color: textColor)),
-                      Text('off',
-                        style: _theme.textTheme.display3
-                          .copyWith(color: textColor))
-                    ],
-                  ))
+                    padding: EdgeInsets.only(top: 16, left: 5),
+                    child: Column(
+                      children: <Widget>[
+                        Text('%', style: _theme.textTheme.headline2.copyWith(color: textColor)),
+                        Text('off', style: _theme.textTheme.headline2.copyWith(color: textColor))
+                      ],
+                    ))
               ])),
           Container(
               padding: EdgeInsets.only(left: AppSizes.sidePadding),
@@ -76,22 +61,17 @@ class OpenFlutterPromoTile extends StatelessWidget {
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: <Widget>[
                     Text(item.title,
-                        style: _theme.textTheme.display1.copyWith(
-                            fontWeight: FontWeight.bold,
-                            color: _theme.primaryColor)),
-                    Text(item.promoCode,
-                        style: _theme.textTheme.body1
-                            .copyWith(color: _theme.primaryColor))
+                        style: _theme.textTheme.headline4
+                            .copyWith(fontWeight: FontWeight.bold, color: _theme.primaryColor)),
+                    Text(item.promoCode, style: _theme.textTheme.bodyText1.copyWith(color: _theme.primaryColor))
                   ])),
           Container(
-              padding: EdgeInsets.only(
-                  left: AppSizes.sidePadding, top: AppSizes.linePadding * 2),
+              padding: EdgeInsets.only(left: AppSizes.sidePadding, top: AppSizes.linePadding * 2),
               width: 155,
               child: Column(children: <Widget>[
                 Text(item.daysLeft.toString() + ' days remaining',
-                    style: _theme.textTheme.display3.copyWith(
-                        fontWeight: FontWeight.bold,
-                        color: _theme.primaryColor)),
+                    style:
+                        _theme.textTheme.headline2.copyWith(fontWeight: FontWeight.bold, color: _theme.primaryColor)),
                 Padding(padding: EdgeInsets.only(bottom: AppSizes.linePadding)),
                 OpenFlutterButton(
                   width: 120,
