@@ -8,11 +8,7 @@ class OpenFlutterInputButton extends StatelessWidget {
   final TextEditingController controller;
 
   const OpenFlutterInputButton(
-      {Key key,
-      @required this.placeHolder,
-      @required this.onClick,
-      @required this.width,
-      @required this.controller})
+      {Key key, @required this.placeHolder, @required this.onClick, @required this.width, @required this.controller})
       : super(key: key);
 
   @override
@@ -27,8 +23,7 @@ class OpenFlutterInputButton extends StatelessWidget {
           Container(
               padding: EdgeInsets.symmetric(horizontal: AppSizes.sidePadding),
               decoration: BoxDecoration(
-                  borderRadius:
-                      BorderRadius.all(Radius.circular(AppSizes.imageRadius)),
+                  borderRadius: BorderRadius.all(Radius.circular(AppSizes.imageRadius)),
                   color: AppColors.white,
                   boxShadow: [
                     BoxShadow(
@@ -40,8 +35,7 @@ class OpenFlutterInputButton extends StatelessWidget {
               width: width - 40,
               child: TextField(
                   controller: controller,
-                  style: _theme.textTheme.display3
-                      .copyWith(color: _theme.primaryColorLight),
+                  style: _theme.textTheme.headline2.copyWith(color: _theme.primaryColorLight),
                   decoration: InputDecoration(
                     border: InputBorder.none,
                     hintText: placeHolder,
