@@ -1,4 +1,3 @@
-
 import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
@@ -6,26 +5,25 @@ import 'package:openflutterecommerce/config/server_addresses.dart';
 import 'package:openflutterecommerce/data/error/exceptions.dart';
 import 'package:openflutterecommerce/domain/usecases/products/products_by_filter_params.dart';
 
-abstract class WoocommercWrapperAbstract{
-   Future<List<dynamic>> getCategoryList({int parentId=0});
-   Future<List<dynamic>> getProductList(ProductsByFilterParams params);
-   Future<List<dynamic>> getPromoList({int userId=0});
+abstract class WoocommercWrapperAbstract {
+  Future<List<dynamic>> getCategoryList({int parentId = 0});
+  Future<List<dynamic>> getProductList(ProductsByFilterParams params);
+  Future<List<dynamic>> getPromoList({int userId = 0});
 }
 
-class WoocommerceWrapper implements WoocommercWrapperAbstract{
-  
+class WoocommerceWrapper implements WoocommercWrapperAbstract {
   final http.Client client;
 
   WoocommerceWrapper({@required this.client});
 
   @override
-  Future<List<dynamic>> getProductList(ProductsByFilterParams params){
+  Future<List<dynamic>> getProductList(ProductsByFilterParams params) {
     //TODO: make remote request using all paramaters
     return _getApiRequest(ServerAddresses.products);
   }
 
   @override
-  Future<List<dynamic>> getCategoryList({int parentId=0}){
+  Future<List<dynamic>> getCategoryList({int parentId = 0}) {
     return _getApiRequest(ServerAddresses.productCategories);
   }
 
@@ -33,9 +31,10 @@ class WoocommerceWrapper implements WoocommercWrapperAbstract{
   Future<List> getPromoList({int userId = 0}) {
     return _getApiRequest(ServerAddresses.promos);
   }
-  
+
   Future<List<dynamic>> _getApiRequest(String url) async {
-    final response = await client.get(url,
+    final response = await client.get(
+      Uri.parse(url),
       headers: {
         'Content-Type': 'application/json',
       },
