@@ -39,15 +39,11 @@ class _Success2ViewState extends State<Success2View> {
             ),
             Padding(
                 padding: EdgeInsets.all(AppSizes.sidePadding * 2),
-                child: Text(
-                    'Your order will be delivered soon. Thank you for choosing our app!',
-                    style: _theme.textTheme.display1)),
+                child: Text('Your order will be delivered soon. Thank you for choosing our app!',
+                    style: _theme.textTheme.headline4)),
             OpenFlutterButton(
               title: 'CONTINUE SHOPPING',
-              onPressed: (() => {
-                    widget.changeView(
-                        changeType: ViewChangeType.Exact, index: 0)
-                  }),
+              onPressed: (() => {widget.changeView(changeType: ViewChangeType.Exact, index: 0)}),
             ),
           ],
         ));
