@@ -31,9 +31,8 @@ class _Main1ViewState extends State<Main1View> {
     var _theme = Theme.of(context);
     var width = MediaQuery.of(context).size.width;
     var widgetWidth = width - AppSizes.sidePadding * 2;
-    return BlocBuilder<HomeBloc, HomeState>(
-      builder: (BuildContext context, HomeState state) {
-        return SingleChildScrollView(
+    return BlocBuilder<HomeBloc, HomeState>(builder: (BuildContext context, HomeState state) {
+      return SingleChildScrollView(
         child: Column(
           children: <Widget>[
             Container(
@@ -56,22 +55,18 @@ class _Main1ViewState extends State<Main1View> {
                       width: width / 2,
                       child: FittedBox(
                         fit: BoxFit.scaleDown,
-                        child: Text(translate('fashionSale'),
-                            style: _theme.textTheme.headline),
+                        child: Text(translate('fashionSale'), style: _theme.textTheme.headline5),
                       ),
                     ),
                     Container(
                       padding: EdgeInsets.only(
-                          left: AppSizes.sidePadding,
-                          bottom: AppSizes.sidePadding,
-                          top: AppSizes.sidePadding),
+                          left: AppSizes.sidePadding, bottom: AppSizes.sidePadding, top: AppSizes.sidePadding),
                       width: 160,
                       child: OpenFlutterButton(
                         title: 'Check',
                         width: 160,
                         height: 48,
-                        onPressed: (() => widget.changeView(
-                            changeType: ViewChangeType.Forward)),
+                        onPressed: (() => widget.changeView(changeType: ViewChangeType.Forward)),
                       ),
                     )
                   ],
@@ -81,29 +76,23 @@ class _Main1ViewState extends State<Main1View> {
               title: 'New',
               linkText: 'View All',
               onLinkTap: () => {
-                Navigator.of(context).pushNamed(OpenFlutterEcommerceRoutes.shop,
-                  arguments: CategoriesParameters(0))
+                Navigator.of(context).pushNamed(OpenFlutterEcommerceRoutes.shop, arguments: CategoriesParameters(0))
               },
               description: 'You’ve never seen it before!',
             ),
             OpenFlutterProductListView(
-              width: widgetWidth, 
+              width: widgetWidth,
               products: widget.products,
-              onFavoritesTap: ( (Product product) => {
-                BlocProvider.of<HomeBloc>(context).add(
-                  HomeAddToFavoriteEvent(
-                    isFavorite: !product.isFavorite,
-                    product: product
-                  )
-                )
-              }),
+              onFavoritesTap: ((Product product) => {
+                    BlocProvider.of<HomeBloc>(context)
+                        .add(HomeAddToFavoriteEvent(isFavorite: !product.isFavorite, product: product))
+                  }),
             ),
             OpenFlutterButton(
               title: 'Next Home Page',
               width: 160,
               height: 48,
-              onPressed: (() =>
-                  widget.changeView(changeType: ViewChangeType.Forward)),
+              onPressed: (() => widget.changeView(changeType: ViewChangeType.Forward)),
             )
           ],
         ),
