@@ -30,8 +30,8 @@ class ProductDetailsView extends StatefulWidget {
       @required this.similarProducts,
       this.category,
       this.hasReviews = false})
-      : assert(product!=null),
-       super(key: key);
+      : assert(product != null),
+        super(key: key);
 
   @override
   _ProductDetailsViewState createState() => _ProductDetailsViewState();
@@ -44,32 +44,29 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
 
   @override
   void initState() {
-    favorite = widget.product?.isFavorite??false;
+    favorite = widget.product?.isFavorite ?? false;
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
     var _theme = Theme.of(context);
-    final dividerTheme =
-        Theme.of(context).copyWith(dividerColor: AppColors.darkGray);
+    final dividerTheme = Theme.of(context).copyWith(dividerColor: AppColors.darkGray);
     var deviceWidth = MediaQuery.of(context).size.width;
     var deviceHeight = MediaQuery.of(context).size.height;
     bloc = BlocProvider.of<ProductBloc>(context);
     return BlocListener(
-        cubit: bloc,
+        bloc: bloc,
         listener: (context, state) {
           if (state is ProductErrorState) {
             return Container(
                 padding: EdgeInsets.all(AppSizes.sidePadding),
-                child: Text('An error occured',
-                    style: _theme.textTheme.display1
-                        .copyWith(color: _theme.errorColor)));
+                child: Text('An error occured', style: _theme.textTheme.headline4.copyWith(color: _theme.errorColor)));
           }
           return Container();
         },
         child: BlocBuilder(
-            cubit: bloc,
+            bloc: bloc,
             builder: (BuildContext context, ProductState state) {
               if (state is ProductLoadedState) {
                 return SingleChildScrollView(
@@ -84,11 +81,8 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
                           height: deviceHeight * 0.52,
                           child: ListView.builder(
                             itemBuilder: (context, index) => Padding(
-                              padding: EdgeInsets.only(right: AppSizes.sidePadding),
-                              child: Image.network(
-                                state.product.images[index].address
-                              )
-                            ),
+                                padding: EdgeInsets.only(right: AppSizes.sidePadding),
+                                child: Image.network(state.product.images[index].address)),
                             scrollDirection: Axis.horizontal,
                             itemCount: state.product.images.length,
                           ),
@@ -99,28 +93,23 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
                           child: Row(
                               mainAxisAlignment: MainAxisAlignment.spaceBetween,
                               mainAxisSize: MainAxisSize.max,
-                              children: 
-                                 (state.product.selectableAttributes != null ?
-                                   state.product.selectableAttributes
-                                      .map((value) => selectionOutlineButton(
-                                          deviceWidth,
-                                          value,
-                                          state.productAttributes.selectedAttributes[value]))
-                                      .toList() : List<Widget>()) +
+                              children: (state.product.selectableAttributes != null
+                                      ? state.product.selectableAttributes
+                                          .map((value) => selectionOutlineButton(
+                                              deviceWidth, value, state.productAttributes.selectedAttributes[value]))
+                                          .toList()
+                                      : List<Widget>()) +
                                   [
                                     OpenFlutterFavouriteButton(
                                       favourite: favorite,
-                                      setFavourite: () => {
-                                        setFavourite(bloc)
-                                      },
+                                      setFavourite: () => {setFavourite(bloc)},
                                     )
-                                  ]) ,
+                                  ]),
                         ),
                         productDetails(_theme),
                         //Function call for Product detail widget
                         Container(
-                          margin: EdgeInsets.only(
-                              left: 16.0, right: 16.0, bottom: 10.0),
+                          margin: EdgeInsets.only(left: 16.0, right: 16.0, bottom: 10.0),
                           child: Text(
                             widget.product.description ?? 'no details',
                             style: TextStyle(fontSize: 15.0),
@@ -158,21 +147,17 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
                           height: 10.0,
                         ),
                         Container(
-                          margin: EdgeInsets.only(
-                              left: 16.0, right: 16.0, bottom: 12.0),
+                          margin: EdgeInsets.only(left: 16.0, right: 16.0, bottom: 12.0),
                           child: Row(
                             mainAxisAlignment: MainAxisAlignment.spaceBetween,
                             crossAxisAlignment: CrossAxisAlignment.center,
                             children: <Widget>[
                               Text(
                                 'You can also like this',
-                                style: TextStyle(
-                                    fontSize: 18.0,
-                                    fontWeight: FontWeight.w600),
+                                style: TextStyle(fontSize: 18.0, fontWeight: FontWeight.w600),
                               ),
                               Text(
-                                widget.similarProducts.length.toString() +
-                                    ' items',
+                                widget.similarProducts.length.toString() + ' items',
                                 style: TextStyle(color: AppColors.lightGray),
                               )
                             ],
@@ -181,14 +166,10 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
                         OpenFlutterProductListView(
                           width: deviceWidth,
                           products: widget.similarProducts,
-                          onFavoritesTap: ( (Product product) => {
-                            BlocProvider.of<HomeBloc>(context).add(
-                              HomeAddToFavoriteEvent(
-                                isFavorite: !product.isFavorite,
-                                product: product
-                              )
-                            )
-                          }),
+                          onFavoritesTap: ((Product product) => {
+                                BlocProvider.of<HomeBloc>(context)
+                                    .add(HomeAddToFavoriteEvent(isFavorite: !product.isFavorite, product: product))
+                              }),
                         )
                       ],
                     ),
@@ -199,9 +180,7 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
             }));
   }
 
-  void setFavourite(
-    ProductBloc bloc
-  ) {
+  void setFavourite(ProductBloc bloc) {
     if (!favorite) {
       bloc.add(ProductAddToFavoritesEvent()); //TODO ask for real parameters if required
     } else {
@@ -212,39 +191,25 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
     });
   }
 
-  void _showSelectAttributeBottomSheet(
-      BuildContext context, 
-      ProductAttribute attribute,
+  void _showSelectAttributeBottomSheet(BuildContext context, ProductAttribute attribute,
       {Function onSelect, String selectedValue}) {
     showModalBottomSheet(
         context: context,
         isScrollControlled: true,
         shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.only(
-            topLeft: Radius.circular(34), topRight: Radius.circular(34)),
+          borderRadius: BorderRadius.only(topLeft: Radius.circular(34), topRight: Radius.circular(34)),
         ),
         builder: (BuildContext context) => AttributeBottomSheet(
-          productAttribute: attribute,
-          selectedValue: selectedValue,
-          onValueSelect: ((String value, ProductAttribute productAttribute)=>
-            {
-              bloc..add(
-                ProductSetAttributeEvent(
-                  value, productAttribute)),
-              Navigator.pop(context),
-              onSelect()
-            }
-          ) 
-        ));
+            productAttribute: attribute,
+            selectedValue: selectedValue,
+            onValueSelect: ((String value, ProductAttribute productAttribute) =>
+                {bloc..add(ProductSetAttributeEvent(value, productAttribute)), Navigator.pop(context), onSelect()})));
   } //modelBottomSheet for selecting size
 
-  Widget selectionOutlineButton(var deviceWidth, ProductAttribute attribute,
-      String alreadySelectedValue) {
+  Widget selectionOutlineButton(var deviceWidth, ProductAttribute attribute, String alreadySelectedValue) {
     //select size and select color widget
     return OutlineButton(
-      onPressed: () => _showSelectAttributeBottomSheet(
-          context, attribute, 
-        selectedValue: alreadySelectedValue),
+      onPressed: () => _showSelectAttributeBottomSheet(context, attribute, selectedValue: alreadySelectedValue),
       child: Container(
         margin: EdgeInsets.only(top: 10.0, bottom: 10.0),
         child: Row(
@@ -252,10 +217,7 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
           children: <Widget>[
             Text(
               alreadySelectedValue ?? attribute.name,
-              style: TextStyle(
-                fontSize: 14,
-                color: AppColors.black,
-                fontWeight: FontWeight.w300),
+              style: TextStyle(fontSize: 14, color: AppColors.black, fontWeight: FontWeight.w300),
             ),
             Icon(Icons.keyboard_arrow_down)
           ],
@@ -267,8 +229,7 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
       focusColor: AppColors.white,
       highlightColor: Colors.white,
       hoverColor: AppColors.red,
-      shape:
-        ContinuousRectangleBorder(borderRadius: BorderRadius.circular(15.0)),
+      shape: ContinuousRectangleBorder(borderRadius: BorderRadius.circular(15.0)),
     );
   }
 
@@ -284,11 +245,11 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
             children: <Widget>[
               Text(
                 widget.product.title,
-                style: theme.textTheme.title,
+                style: theme.textTheme.headline6,
               ),
               Text(
                 '\$' + widget.product.price.toString(),
-                style: theme.textTheme.title,
+                style: theme.textTheme.headline6,
               )
             ],
           ),
@@ -296,7 +257,7 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
               ? Container()
               : Text(
                   widget.category.name,
-                  style: theme.textTheme.body1,
+                  style: theme.textTheme.bodyText1,
                 ),
           SizedBox(
             height: 5,
@@ -329,8 +290,7 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
   }
 
   void _addItemToCart(BuildContext context, ProductLoadedState state) async {
-    if (state.productAttributes.selectedAttributes.length ==
-        state.product.selectableAttributes.length) {
+    if (state.productAttributes.selectedAttributes.length == state.product.selectableAttributes.length) {
       BlocProvider.of<ProductBloc>(context).add(ProductAddToCartEvent());
       await Navigator.pushNamed(context, OpenFlutterEcommerceRoutes.cart);
     } else {
@@ -338,13 +298,12 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
         final attribute = state.product.selectableAttributes[i];
         if (!state.productAttributes.selectedAttributes.containsKey(attribute)) {
           await _showSelectAttributeBottomSheet(context, attribute,
-            onSelect: i == 0 ?
-            (() => {
-              BlocProvider.of<ProductBloc>(context).add(ProductAddToCartEvent()),
-              Navigator.pushNamed(context, OpenFlutterEcommerceRoutes.cart)
-            })
-            : null
-          );
+              onSelect: i == 0
+                  ? (() => {
+                        BlocProvider.of<ProductBloc>(context).add(ProductAddToCartEvent()),
+                        Navigator.pushNamed(context, OpenFlutterEcommerceRoutes.cart)
+                      })
+                  : null);
         }
       }
     }
