@@ -27,12 +27,10 @@ class _ProfileScreenState extends State<ProfileScreen> {
       child: OpenFlutterScaffold(
         title: 'My Profile',
         body: BlocProvider<ProfileBloc>(
-          create: (context) {
-            return ProfileBloc(
-            )
-              ..add(ProfileStartEvent());
-          },
-          child: ProfileWrapper()),
+            create: (context) {
+              return ProfileBloc()..add(ProfileStartEvent());
+            },
+            child: ProfileWrapper()),
         bottomMenuIndex: 4,
       ),
     );
@@ -48,9 +46,9 @@ class _ProfileWrapperState extends OpenFlutterWrapperState<ProfileWrapper> {
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<ProfileBloc, ProfileState>(
-      cubit: BlocProvider.of<ProfileBloc>(context),
-      builder: (BuildContext context, ProfileState state) {
-        return getPageView(<Widget>[
+        bloc: BlocProvider.of<ProfileBloc>(context),
+        builder: (BuildContext context, ProfileState state) {
+          return getPageView(<Widget>[
             ProfileView(changeView: changePage),
             MyOrdersView(changeView: changePage),
             MyShippingAddressesView(changeView: changePage),
@@ -60,6 +58,6 @@ class _ProfileWrapperState extends OpenFlutterWrapperState<ProfileWrapper> {
             SettingsView(changeView: changePage),
             MyOrderDetailsView(changeView: changePage),
           ]);
-      });
+        });
   }
 }
