@@ -57,7 +57,7 @@ class _SettingsViewState extends State<SettingsView> {
     return BlocProvider<SettingsBloc>(
       create: (context) => settingsBloc,
       child: BlocBuilder<SettingsBloc, SettingsState>(
-          cubit: settingsBloc,
+          bloc: settingsBloc,
           builder: (context, state) {
             _fullNameController.text = state.settings.fullName;
             _dateOfBirthController.text = state.settings.dateOfBirth;
@@ -87,9 +87,8 @@ class _SettingsViewState extends State<SettingsView> {
                     controller: _fullNameController,
                     hint: 'Full Name',
                     horizontalPadding: 0,
-                    onValueChanged: (value) => settingsBloc
-                      ..add(UpdateFullNameEvent(
-                          fullName: value.toString().trim())),
+                    onValueChanged: (value) =>
+                        settingsBloc..add(UpdateFullNameEvent(fullName: value.toString().trim())),
                   ),
                   SizedBox(
                     height: 24,
@@ -98,9 +97,8 @@ class _SettingsViewState extends State<SettingsView> {
                     controller: _dateOfBirthController,
                     hint: 'Date of Birth',
                     horizontalPadding: 0,
-                    onValueChanged: (value) => settingsBloc.add(
-                        UpdateDateOfBirthEvent(
-                            dateOfBirth: value.toString().trim())),
+                    onValueChanged: (value) =>
+                        settingsBloc.add(UpdateDateOfBirthEvent(dateOfBirth: value.toString().trim())),
                   ),
                   SizedBox(
                     height: 55,
@@ -110,8 +108,7 @@ class _SettingsViewState extends State<SettingsView> {
                     children: <Widget>[
                       Text(
                         'Password',
-                        style: TextStyle(
-                            fontSize: 16, fontWeight: FontWeight.w500),
+                        style: TextStyle(fontSize: 16, fontWeight: FontWeight.w500),
                       ),
                       GestureDetector(
                         onTap: () {
@@ -119,10 +116,7 @@ class _SettingsViewState extends State<SettingsView> {
                         },
                         child: Text(
                           'Change',
-                          style: TextStyle(
-                              fontSize: 14,
-                              color: AppColors.lightGray,
-                              fontWeight: FontWeight.w500),
+                          style: TextStyle(fontSize: 14, color: AppColors.lightGray, fontWeight: FontWeight.w500),
                         ),
                       ),
                     ],
@@ -148,8 +142,7 @@ class _SettingsViewState extends State<SettingsView> {
                         //trackColor: AppColors.lightGray,
                         value: state.settings.notifySales,
                         activeColor: AppColors.success,
-                        onChanged: (newValue) => settingsBloc
-                            .add(UpdateNotifySalesEvent(notifySales: newValue)),
+                        onChanged: (newValue) => settingsBloc.add(UpdateNotifySalesEvent(notifySales: newValue)),
                       ),
                     ],
                   ),
@@ -167,9 +160,7 @@ class _SettingsViewState extends State<SettingsView> {
                         //trackColor: AppColors.lightGray,
                         value: state.settings.notifyArrivals,
                         activeColor: AppColors.success,
-                        onChanged: (newValue) => settingsBloc
-                          ..add(UpdateNotifyArrivalsEvent(
-                              notifyArrivals: newValue)),
+                        onChanged: (newValue) => settingsBloc..add(UpdateNotifyArrivalsEvent(notifyArrivals: newValue)),
                       ),
                     ],
                   ),
@@ -187,9 +178,7 @@ class _SettingsViewState extends State<SettingsView> {
                         //trackColor: AppColors.lightGray,
                         value: state.settings.notifyDelivery,
                         activeColor: AppColors.success,
-                        onChanged: (newValue) => settingsBloc
-                          ..add(UpdateNotifyDeliveryEvent(
-                              notifyDelivery: newValue)),
+                        onChanged: (newValue) => settingsBloc..add(UpdateNotifyDeliveryEvent(notifyDelivery: newValue)),
                       ),
                     ],
                   ),
@@ -209,21 +198,19 @@ class _SettingsViewState extends State<SettingsView> {
     showModalBottomSheet(
         context: context,
         shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.only(
-              topLeft: Radius.circular(34), topRight: Radius.circular(34)),
+          borderRadius: BorderRadius.only(topLeft: Radius.circular(34), topRight: Radius.circular(34)),
         ),
         builder: (BuildContext context) => BlocProvider<PasswordBloc>(
               create: (context) => passwordBloc,
               child: BlocBuilder<PasswordBloc, PasswordState>(
-                  cubit: passwordBloc,
+                  bloc: passwordBloc,
                   builder: (context, state) {
                     if (state is PasswordChangedState) {
                       WidgetsBinding.instance.addPostFrameCallback((_) {
                         // close bottom sheet
                         Navigator.pop(context);
 
-                        _showAlertDialog(context, 'Success',
-                            'Password changed successfully');
+                        _showAlertDialog(context, 'Success', 'Password changed successfully');
 
                         clearPasswordFields();
                       });
@@ -238,9 +225,7 @@ class _SettingsViewState extends State<SettingsView> {
                       padding: AppSizes.bottomSheetPadding,
                       decoration: BoxDecoration(
                           color: AppColors.background,
-                          borderRadius: BorderRadius.only(
-                              topLeft: Radius.circular(34),
-                              topRight: Radius.circular(34)),
+                          borderRadius: BorderRadius.only(topLeft: Radius.circular(34), topRight: Radius.circular(34)),
                           boxShadow: []),
                       child: SingleChildScrollView(
                         child: Column(
@@ -258,10 +243,7 @@ class _SettingsViewState extends State<SettingsView> {
                             ),
                             Text(
                               'Password Change',
-                              style: TextStyle(
-                                  color: AppColors.black,
-                                  fontSize: 18,
-                                  fontWeight: FontWeight.bold),
+                              style: TextStyle(color: AppColors.black, fontSize: 18, fontWeight: FontWeight.bold),
                             ),
                             SizedBox(
                               height: 18,
@@ -319,14 +301,9 @@ class _SettingsViewState extends State<SettingsView> {
                                 height: 48,
                                 onPressed: () => passwordBloc
                                   ..add(ChangePasswordEvent(
-                                      currentPassword:
-                                          _currentPasswordController.text
-                                              .trim(),
-                                      newPassword:
-                                          _newPasswordController.text.trim(),
-                                      repeatNewPassword:
-                                          _repeatPasswordController.text
-                                              .trim())))
+                                      currentPassword: _currentPasswordController.text.trim(),
+                                      newPassword: _newPasswordController.text.trim(),
+                                      repeatNewPassword: _repeatPasswordController.text.trim())))
                           ],
                         ),
                       ),
@@ -335,8 +312,7 @@ class _SettingsViewState extends State<SettingsView> {
             ));
   }
 
-  Future<void> _showAlertDialog(
-      BuildContext context, String title, String content) {
+  Future<void> _showAlertDialog(BuildContext context, String title, String content) {
     return showDialog<void>(
       context: context,
       builder: (BuildContext context) {
