@@ -28,15 +28,12 @@ class OpenFlutterClickableLine extends StatelessWidget {
     return InkWell(
         onTap: (() => {onTap(sortRules)}),
         child: Container(
-            padding: EdgeInsets.symmetric(
-                horizontal: AppSizes.sidePadding,
-                vertical: AppSizes.linePadding),
+            padding: EdgeInsets.symmetric(horizontal: AppSizes.sidePadding, vertical: AppSizes.linePadding),
             alignment: Alignment.centerLeft,
             height: height,
             width: width,
             color: backgroundColor,
             child: Text(title,
-                style: _theme.textTheme.display1.copyWith(
-                    fontWeight: FontWeight.normal, color: textColor))));
+                style: _theme.textTheme.headline4.copyWith(fontWeight: FontWeight.normal, color: textColor))));
   }
 }
