@@ -28,10 +28,7 @@ class ErrorDialog extends StatelessWidget {
           Padding(
             padding: EdgeInsets.symmetric(horizontal: 16.0),
             child: Text(mainText,
-                style: Theme.of(context)
-                    .textTheme
-                    .display1
-                    .copyWith(color: Theme.of(context).errorColor)),
+                style: Theme.of(context).textTheme.headline4.copyWith(color: Theme.of(context).errorColor)),
           ),
           SizedBox(
             height: 12.0,
