@@ -48,10 +48,8 @@ class OpenFlutterPaymentCardPreview extends StatelessWidget {
                   width: width - AppSizes.sidePadding,
                   height: 38,
                   child: Text(cardNumber,
-                      style: _theme.textTheme.title.copyWith(
-                          fontSize: 24,
-                          color: AppColors.white,
-                          letterSpacing: 6)),
+                      style:
+                          _theme.textTheme.headline6.copyWith(fontSize: 24, color: AppColors.white, letterSpacing: 6)),
                 ),
                 Positioned(
                   top: 210,
@@ -59,7 +57,7 @@ class OpenFlutterPaymentCardPreview extends StatelessWidget {
                   width: width - AppSizes.sidePadding,
                   height: 38,
                   child: Text(cardHolderName,
-                      style: _theme.textTheme.display3.copyWith(
+                      style: _theme.textTheme.headline2.copyWith(
                           //fontSize: 34,
                           color: AppColors.white)),
                 ),
@@ -68,13 +66,11 @@ class OpenFlutterPaymentCardPreview extends StatelessWidget {
                   left: 185,
                   width: width - AppSizes.sidePadding,
                   height: 38,
-                  child: Text(
-                      expirationMonth.toString().padLeft(2, '0') +
-                          '/' +
-                          expirationYear.toString().padLeft(2, '0'),
-                      style: _theme.textTheme.display3.copyWith(
-                          //fontSize: 34,
-                          color: AppColors.white)),
+                  child:
+                      Text(expirationMonth.toString().padLeft(2, '0') + '/' + expirationYear.toString().padLeft(2, '0'),
+                          style: _theme.textTheme.headline2.copyWith(
+                              //fontSize: 34,
+                              color: AppColors.white)),
                 ),
               ],
             )));
