@@ -46,7 +46,7 @@ class OpenFlutterViewOptions extends StatelessWidget {
                   Icon(Icons.filter_list),
                   Padding(
                     padding: EdgeInsets.only(left: 5),
-                    child: Text('Filters', style: _theme.textTheme.body1),
+                    child: Text('Filters', style: _theme.textTheme.bodyText1),
                   )
                 ])),
             InkWell(
@@ -57,8 +57,7 @@ class OpenFlutterViewOptions extends StatelessWidget {
                   Icon(Icons.import_export),
                   Container(
                     padding: EdgeInsets.only(left: 5),
-                    child: Text(sortRules?.getSortTitle() ?? '',
-                        style: _theme.textTheme.body1),
+                    child: Text(sortRules?.getSortTitle() ?? '', style: _theme.textTheme.bodyText1),
                   )
                 ])),
             IconButton(
@@ -75,83 +74,69 @@ class OpenFlutterViewOptions extends StatelessWidget {
   }
 
   void _showFilterWindow(BuildContext context) {
-    Navigator.of(context)
-        .pushNamed(OpenFlutterEcommerceRoutes.filters, arguments: filterRules);
+    Navigator.of(context).pushNamed(OpenFlutterEcommerceRoutes.filters, arguments: filterRules);
   }
 
   void _showSortOptions(BuildContext context) {
     showModalBottomSheet(
         context: context,
         shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.only(
-            topLeft: Radius.circular(34.0),
-            topRight: Radius.circular(34.0)
-          ),
+          borderRadius: BorderRadius.only(topLeft: Radius.circular(34.0), topRight: Radius.circular(34.0)),
         ),
         backgroundColor: Colors.white,
         builder: (context) {
-          return Column(
-            children: <Widget>[
-              Container(
-                padding: EdgeInsets.all(AppSizes.sidePadding),
-                child: Container(
-                  width: 60,
-                  height: 6,
-                  decoration: BoxDecoration(
-                    color: AppColors.darkGray,
-                    borderRadius: BorderRadius.circular(AppSizes.imageRadius),
-                  ),
+          return Column(children: <Widget>[
+            Container(
+              padding: EdgeInsets.all(AppSizes.sidePadding),
+              child: Container(
+                width: 60,
+                height: 6,
+                decoration: BoxDecoration(
+                  color: AppColors.darkGray,
+                  borderRadius: BorderRadius.circular(AppSizes.imageRadius),
                 ),
               ),
-              Text('Sort by',
-                style: Theme.of(context).textTheme.display1),
-              Padding(padding: EdgeInsets.only(top: AppSizes.sidePadding),),
-              ...sortRules.sortTextVariants
+            ),
+            Text('Sort by', style: Theme.of(context).textTheme.headline4),
+            Padding(
+              padding: EdgeInsets.only(top: AppSizes.sidePadding),
+            ),
+            ...sortRules.sortTextVariants
                 .map((key, value) => MapEntry(
                       key,
                       Container(
-                        padding: EdgeInsets.symmetric(
-                            horizontal: AppSizes.sidePadding,
-                            vertical: AppSizes.linePadding),
+                        padding: EdgeInsets.symmetric(horizontal: AppSizes.sidePadding, vertical: AppSizes.linePadding),
                         alignment: Alignment.centerLeft,
-                        color: sortRules.sortType == key
-                            ? AppColors.red
-                            : AppColors.white,
+                        color: sortRules.sortType == key ? AppColors.red : AppColors.white,
                         child: Row(
                           children: <Widget>[
                             Expanded(
                               child: InkWell(
                                 child: Text(value,
-                                  style: Theme.of(context).textTheme.display1.copyWith(        
-                                      fontWeight: FontWeight.normal,
-                                      color: sortRules.sortType == key
-                                        ? AppColors.white
-                                        : AppColors.black)),
+                                    style: Theme.of(context).textTheme.headline4.copyWith(
+                                        fontWeight: FontWeight.normal,
+                                        color: sortRules.sortType == key ? AppColors.white : AppColors.black)),
                                 onTap: () {
                                   onSortChanged(SortRules(
-                                      sortOrder:
-                                        sortRules.sortType == key ?
-                                        (
-                                          sortRules.sortOrder == SortOrder.FromLowestToHighest ?
-                                          SortOrder.FromHighestToLowest : SortOrder.FromLowestToHighest
-                                        ) :
-                                        sortRules.sortOrder,
+                                      sortOrder: sortRules.sortType == key
+                                          ? (sortRules.sortOrder == SortOrder.FromLowestToHighest
+                                              ? SortOrder.FromHighestToLowest
+                                              : SortOrder.FromLowestToHighest)
+                                          : sortRules.sortOrder,
                                       sortType: key));
                                   Navigator.pop(context);
                                 },
                               ),
                             ),
                             IconButton(
-                              icon: Icon(
-                                sortRules.sortOrder == SortOrder.FromHighestToLowest ?
-                                FontAwesomeIcons.sortAlphaUp :
-                                FontAwesomeIcons.sortAlphaDown),
-                              color: sortRules.sortType == key ?
-                                Theme.of(context).primaryColor 
-                                : Theme.of(context).backgroundColor,
+                              icon: Icon(sortRules.sortOrder == SortOrder.FromHighestToLowest
+                                  ? FontAwesomeIcons.sortAlphaUp
+                                  : FontAwesomeIcons.sortAlphaDown),
+                              color: sortRules.sortType == key
+                                  ? Theme.of(context).primaryColor
+                                  : Theme.of(context).backgroundColor,
                               onPressed: () {
-                                onSortChanged(
-                                    sortRules.copyWithChangedOrder());
+                                onSortChanged(sortRules.copyWithChangedOrder());
                                 Navigator.pop(context);
                               },
                             )
@@ -161,7 +146,7 @@ class OpenFlutterViewOptions extends StatelessWidget {
                     ))
                 .values
                 .toList(growable: false),
-            ]);
+          ]);
         });
   }
 }
